+++ cterasdk/asynchronous/core/remote.py
@@ -1,17 +1,17 @@
-from .enum import DeviceType
-from ..objects.synchronous import edge, drive
-from ..common import parse_base_object_ref
+from ...core.enum import DeviceType
+from ...objects.asynchronous import edge, drive
+from ...common import parse_base_object_ref
 
 
 def remote_command(core, device):
     tenant = parse_base_object_ref(device.portal).name
-    base = f'{core.ctera.baseurl}/devicecmdnew/{tenant}/{device.name}'
+    base = f'{core.v1.ctera.baseurl}/devicecmdnew/{tenant}/{device.name}'
 
     ManagedDevice = None
     if device.deviceType in DeviceType.Gateways:
-        ManagedDevice = edge.Edge(core=core, base=base)
+        ManagedDevice = edge.AsyncEdge(core=core, base=base)
     elif device.deviceType in DeviceType.Agents:
-        ManagedDevice = drive.Drive(core=core, base=base)
+        ManagedDevice = drive.AsyncDrive(core=core, base=base)
     elif device.deviceType == "Mobile":
         return device
     else: